ml-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
